# packages/api-client/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/environments/{environmentId}/services": {
			"get": {
				"operationId": "environmentServicesList",
				"parameters": [
					{
						"name": "environmentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ServiceViewDto"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List services in an environment",
				"tags": ["services"]
			},
			"post": {
				"operationId": "environmentServicesCreate",
				"parameters": [
					{
						"name": "environmentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateServiceDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Create a service in an environment",
				"tags": ["services"]
			}
		},
		"/api/services/{serviceId}": {
			"get": {
				"operationId": "servicesGet",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get service details",
				"tags": ["services"]
			},
			"patch": {
				"operationId": "servicesUpdate",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateServiceDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceViewDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Update a service",
				"tags": ["services"]
			},
			"delete": {
				"operationId": "servicesDelete",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceOkDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Delete a service",
				"tags": ["services"]
			}
		},
		"/api/services/{serviceId}/connection": {
			"get": {
				"operationId": "servicesConnectionGet",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceConnectionDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get managed database connection details",
				"tags": ["services"]
			}
		},
		"/api/services/{serviceId}/status": {
			"get": {
				"operationId": "serviceStatusGet",
				"parameters": [
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceRuntimeDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get live service runtime status",
				"tags": ["services"]
			}
		},
		"/api/environments/{environmentId}/services/status": {
			"get": {
				"operationId": "environmentServiceStatusList",
				"parameters": [
					{
						"name": "environmentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ServiceRuntimeEntryDto"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List live runtime status for services in an environment",
				"tags": ["services"]
			}
		},
		"/api/services/{serviceId}/logs": {
			"get": {
				"operationId": "serviceLogsGet",
				"parameters": [
					{
						"name": "tailLines",
						"required": false,
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "pod",
						"required": false,
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "serviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceLogsDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Get recent container logs for a service",
				"tags": ["services"]
			}
		},
...
	},
	"info": {
		"title": "kubwave API",
		"description": "",
		"version": "dev",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
...
			"AutoDeployViewDto": {
				"type": "object",
				"properties": {
					"enabled": {
						"type": "boolean"
					},
					"lastPolledCommit": {
						"type": "string",
						"nullable": true
					},
					"lastPolledAt": {
						"type": "string",
						"nullable": true
					},
					"nextPollAt": {
						"type": "string",
						"nullable": true
					},
					"lastPollError": {
						"type": "string",
						"nullable": true
					}
				},
				"required": ["enabled", "lastPolledCommit", "lastPolledAt", "nextPollAt", "lastPollError"]
			},
			"ServiceViewDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"environmentId": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": ["docker-image", "dockerfile", "public-repo", "private-repo", "postgres", "mysql", "mariadb", "mongodb"]
					},
					"config": {
						"type": "object",
						"additionalProperties": true
					},
					"autoDeploy": {
						"$ref": "#/components/schemas/AutoDeployViewDto"
					},
					"internalDomain": {
						"type": "string",
						"nullable": true
					},
					"defaultUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"environmentId",
					"name",
					"description",
					"type",
					"config",
					"autoDeploy",
					"internalDomain",
					"defaultUrl",
					"createdAt",
					"updatedAt"
				]
			},
			"AutoDeployInputDto": {
				"type": "object",
				"properties": {
					"enabled": {
						"type": "boolean",
						"default": false
					}
				},
				"required": ["enabled"]
			},
			"CreateServiceDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					},
					"description": {
						"type": "string",
						"maxLength": 1000
					},
					"type": {
						"type": "string",
						"enum": ["docker-image", "dockerfile", "public-repo", "private-repo", "postgres", "mysql", "mariadb", "mongodb"]
					},
					"config": {
						"type": "object",
						"additionalProperties": true
					},
					"autoDeploy": {
						"$ref": "#/components/schemas/AutoDeployInputDto"
					}
				},
				"required": ["name", "type", "config"]
			},
			"UpdateServiceDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					},
					"description": {
						"type": "string",
						"maxLength": 1000
					},
					"config": {
						"type": "object",
						"additionalProperties": true
					},
					"autoDeploy": {
						"$ref": "#/components/schemas/AutoDeployInputDto"
					}
				}
			},
			"ServiceOkDto": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": ["ok"]
			},
			"ServiceConnectionDto": {
				"type": "object",
				"properties": {
					"engine": {
						"type": "string",
						"enum": ["postgres", "mysql", "mariadb", "mongodb"]
					},
					"host": {
						"type": "string"
					},
					"port": {
						"type": "number"
					},
					"username": {
						"type": "string"
					},
					"database": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"uri": {
						"type": "string"
					}
				},
				"required": ["engine", "host", "port", "username", "database", "password", "uri"]
			},
			"ServiceRuntimeDto": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"enum": ["running", "degraded", "progressing", "stopped", "failed", "not_deployed", "unknown"]
					},
					"readyReplicas": {
						"type": "number"
					},
					"desiredReplicas": {
						"type": "number"
					},
					"updatedReplicas": {
						"type": "number"
					},
					"availableReplicas": {
						"type": "number"
					}
				},
				"required": ["status", "readyReplicas", "desiredReplicas", "updatedReplicas", "availableReplicas"]
			},
			"ServiceRuntimeEntryDto": {
				"type": "object",
				"properties": {
					"serviceId": {
						"type": "string",
						"format": "uuid"
					},
					"runtime": {
						"$ref": "#/components/schemas/ServiceRuntimeDto"
					}
				},
				"required": ["serviceId", "runtime"]
			},
			"ServiceLogEntryDto": {
				"type": "object",
				"properties": {
					"pod": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": ["pod", "timestamp", "message"]
			},
			"ServiceLogsDto": {
				"type": "object",
				"properties": {
					"available": {
						"type": "boolean"
					},
					"pods": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"entries": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ServiceLogEntryDto"
						}
					}
				},
				"required": ["available", "pods", "entries"]
			},
...
		}
	}
}